[#1792] Make the FTS index method configurable
GiST indexes are faster to create and take less memory, but are about 3 times
slower than GIN indexes. Depending on the use-case, one is better than the
other. As we can't foresee all ways CKAN will be used, with this patch we allow
the sysadmin to pick what suits best for her.