Merge IResourceModification into IResourceController
This is done for consistency in the code (this makes it similar
to IPackageController).

It is worth noting that IResourceController already had a
before_show method which is called in the other methods as well
and some parameters could therefore be redundant but they are left
intact because it becomes difficult to know what is what if we ask
devs to use before_show in order to retrieve a variable they want
to use (and before_show can actually be called multiple times in
some cases).