Allows use of CKAN_INI env-var to locate config
Implements #1597

If you don't specify a config file using the -c option, CKAN defaults to
looking for development.ini in the current directory. This patch will
first check the -c parameter, and if it doesn't find it will use the
content of the CKAN_INI environment variable. If it cannot find that,
it will then default to development.ini in the current directory.