[#1852] Create default views on dataset creation and update
Rather than relying on system plugins implementing `after_update` and
`after_create` to create default views, we now call the necessary
functions in `datapreview` to create them depeding on the default views
configuration setting.

The resources on the dataset dict that need to be passed to this
function on `package_create` and `package_update` didn't originally have
the ids, as the commit hadn't take place yet. We now do a flush just
before passing the dicts and add the ids so they can be used to create
the resource views.

There is no need to specifically create the default views on
`resource_create` and `resource_update`, as these call `package_update`
anyway.