[#1852] Allow some views to be always available
Some views like the image and website one were always available as an
option when creating a new view against a resource. With the new
handling of default views this was no longer possible, so to support
this use case a new `always_available` key has been added to the plugin
info.