Improve the efficiency of user_show
Instead of returning up to 50 of the datasets that a user can administer
now the call must specify that it wants those datasets by passing in
```include_datasets:True```.

The num_followers value must also be explicitly requested in order to be
included in the resulting dict.

Revisions are no longer returned, users should call user_activity_list
instead.

fixes #2130