[#2142] Fix crash on creating datasets in the web UI
Try to create any dataset in the web UI and after submitting the first
form it crashes with logic.NotFound.

The package controller is passing a package dict to the
resource_create() auth function, which is surely wrong.

Change it to pass a dict with just the "package_id" for the (not created
yet) new resource instead, this seems to be all the resource_create()
auth function needs.