Fixes broken test on master when celery is installed.
It appears that test_40_task_resource_status was never run if you do not
have celery installed, and so didn't show up on travis.  The fix was to
install celery, and to wrap the 'resource_status_show' with the correct
url as a string.  This threw up another problem in that the results from
the resource status check did not match what the tests assumed they did.

Either this has changed since it last caused a problem for someone, or
it has been hidden for a very long time by the SkipTest :(