Fixes #2210 - changing group_create and group_list calls
Instead of calling group_dictize and returning the results at the end of
group_create this now returns group_show instead for consistency.  This
ensures the show schema is used to validate and process the result
returned to the user.  It should make it easier to call create, modify
the response and then post to update.

Instead of using group_list_dictize the group_list call now uses
group_show to return the relevant data.

The deprecated use of packages as either a list or a count has been
removed as package_count is already returned on each group_show call.

This is more than likely a breaking change and will need close checks to
make sure it is accurate.