[#2242] Add recline_view to the default view types
As part of #2140 DataProxy was reenabled, so to keep the old behaviour
of showing a Recline data explorer when uploading or linking to a
csv/excel we added it to the default view types. The data explorer view
plugin no longer requires the DataStore, as it can use DataProxy.

Also tweaked the client side error message to reflect if the error came
from the DataStore or the DataProxy.