Add command to run when installing from source
I came to this page from the "install CKAN from source" page. The set permissions part did not work for me - I replaced sudo ckan with paster --plugin=ckan but that was not enough.

I took the command from http://docs.ckan.org/en/ckan-2.0/datastore-setup.html and that one worked - included the "postgres" keyword and configuration file path. It would be helpful to include it directly to the note dedicated to the ones installing from source.