Fixes #2412 by checking for less correct 400 response
Although the code checks for Error 405 which would be the correct response from a server when HEAD is not supported, I have come across a server where 400 is issued with the text "Invalid method in request", which essentially seems to mean the same thing. If we check for this too we can save problems with annoying servers.