[#2429] Make the system info table revisioned
If we are going to allow the configuration to be changed remotely it
will be useful to be able to track changes on it (when was it done,
which user did it, etc.) The best way we have now for doing is
revisioning, even though it's a pain to set up, as demonstrated by this
commit. Fields need to be added manually and the model definition is
much more complex.

Added preliminary tests.