[#2429] config_option_show action.
Sysadmins can call this action with a key (eg. 'ckan.site_title') to
retrieve a config value. The value is returned either from the
system_info table of the database, or secondly, from the config object
as defined in the ini file. Only config options in the schema
whitelisted as editable will be returned. Non-editable options, or non
existing options, will return a ValidationError.