[#2429] Update admin controller to use new actions
Bring the config method of the admin controller more in line with how
the rest of form/controllers work in ckan. Call `config_option_update`
action, catch and display ValidationErrors.

Pass all editable options to the form template, even if the fields are
not displayed by default. This way extensions can add fields to the form
and these will get stored.

Added an example on how to extend the template to the example_iconfigurer
extension.