[#2553] Fix autodetect for tsv resources
When you upload or link to a TSV file and don't specify a resource format,
the format that you end up with is text/tab-separated-values.
This is not recognized by the DataPusher and the resource is not uploaded to
 the DataStore.

This patch adds this alternative representation to the canonical resource
format list.