Introduces IDataPusher plugin
This new plugin allows users to reject resources for submission,
and also to be notified once a datapush is complete.

The two methods in the interface are:

can_upload: which will abort datapusher_submit calls if they return
True.
after_upload: Called with the resource ID after the upload into the
datastore is complete.