[#2554] Refactor group_list to only query necessary fields
Refactor organization/group_list to only query the necessary fields by default
(id, name, title, package_count) depending on the required sort. This massively
speeds up the default query without all_fields. In part this is because then we
no longer need to get all fields in all groups on all cases to do the sorting.
There is a minor drawback in that then we can't take private datasets into
account when sorting by number of datasets. The actual number displayed will
take private datasets into account, as this comes from the dictization, but
there might be inconsistencies (note that the "order by datasets option" is
not offered by default on the UI)