[#2234] Improve datastore connection checks on migration
If the datastore connection is set on the config file but the user or
database don't actually exist, skip the migration.

Note that this only happens when the config options are there but the
datastore plugin is not loaded (like in many extensions that extend
their test.ini with the core test-core.ini). Otherwise the connection
will crash anyway during the loading of the plugin itself.