Adds metadata_created as a proper DB field.
Currently the metadata_created is calculated on each access by querying
the package_revision table and finding the oldest revision.  There's
little advantage to this over setting the field to be a column on the
package table and defaulting to now().  This means it can be overwritten
by those components, like harvesting, that require the ability to do so
- without meddling with revisions.

This should fix #2735