Replace markdown library with python-markdown
Replaces the previous markdown renderer that was problematic (and
unmaintained) with the more widely used python-markdown.

As python-markdown has deprecated the safe_mode for removing HTML, this
PR also contains bleach to strip out any HTML provided - as recommended
by the python-markdown docs.

The only noticeable change in rendering is that the previous library
inserted a newline before a closing p tag, which python-markdown doesn't
do.  This is unlikely to change how the content is rendered.