Fix app lookup
app_name should be the name (ie `pylons_app`) not the actual app

This issue arised because currently the functional tests are broken and
don't call the `can_handle_url` method of the partyline controller. This
means that answers was an empty list and the default option was being
used (crashing because the app lookup failed).

The issue with the functional tests is that when running the tests with
`--with-pylons` option, Pylons will create a WSGI app:

venv/lib/python2.7/site-packages/pylons/test.py:73

Aparently this is to load the configuration. The WSGI Party invites are
sent during the initialization of this first app.

When running a functional test and calling `app = self._get_test_app()`
a new app is created, and invites are sent again. But at this point the
`partyline_connected` attribute has already been set up on the Pylons
globals object, so the URL handler won't be registered.

        if hasattr(g, 'partyline_connected'):
            base.abort(404)

This means that when running the functional tests, only the Flask url
handler is currently called.