[#2485] Enforce order of precedence on request handler middleware
Enforce Flask Extension > Pylons Extension > Flask Core > Pylons Core

Pylons requests are marked core=True or core=False depending on whether
they were set up in core or by an extension. This information is passed
to the handler so it can prioritize which app to use.

TODO: It's still not clear how we will mark Flask routes as core or not,
it will depend on how we end up setting up routing