[#2845] Improve origin detection on Pylons mapper
We can not add custom params to the defaults dict because they get added
to the URL built, eg url_for('group_index') returns `/group?core=True

Ideally we want to keep this flag attached to the actual route, because
it's hard to come up with unique keys for routes which are not named to
keep them on a separate registry. As we can get the actual matched Route
object using `map.routematch` we add a `_ckan_core` attribute to it when
building the mapper.