[#2661] Add revision if owner-org updated.
package_owner_org_update is called during a package_update to handle
updating the owner organization. A package revision is created during
this by package_update. If package_owner_org_update is called in
isolation, no revision is created and an error occurs in vdm. This
commit ensures a revision is created when package_owner_org_update is
called outside of a package_update.