[#3148] Remove WSGI Party
When the App Dispatcher middleware was implemented on #2905 we used WSGI Party
to allow communication between the Pylons and Flask applications. After having
worked extensively on the Flask POC branches this has turned out to be
unnecessary and to add quite a lot of complexity, specially with the language
used around parties, invites, etc.

To set it up there were two internal requests that were done while the app was
being initialized, which were really confusing when trying to debug things. We
didn't find any use case for having the two applications actually talk to each
other, and the asking each app if they can handle the incoming request can be
done just by calling a method directly.

This commit removes all WSGI Party logic and its requirement, and extends the
Flask and PylonsApp app objects with our own can_handle_request(environ)
method, which is directly called by AskAppDispatcherMiddleware on each request.