[#2952] Depcreate `db load` and `db dump`
Deprecates the paster commands `db load` and `db dump` in favor of
PostgreSQL's `pg_dump` and `pg_restore`. `db load` and `db dump` are
kept but print a warning, and their documentation is replaced with a
documentation of how to dump/import the database using `pg_dump` and
`pg_restore`.