[#2842] New CKAN config object
Rather than rely on the Pylons (or Flask) config object we define our
own in ckan.common. This is a dict-like object (so fully backwards compatible)
that also proxies any changes to the Flask and Pylons configuration objects
(if they are available)

This should be the only configuration object used in all code unless we really
need to access the underlying Flask or Pylons objects for some reason.

The `ckan.common.config` instance is initialized in the `load_environment`
method with the values of the ini file or env vars.

This is actually a proxy to a property from a Werkzeug Local object, meaning
that `config` is thread-local safe, like its Flask and Pylons counterparts.

See http://werkzeug.pocoo.org/docs/0.11/local/

I tried to separate all Pylons specific stuff (things like `pylons.paths`,
`routes.map` etc) to Pylons own config object to keep the main config object
clean but it proved too difficult, not only because we access these keys from
different parts of the code (which can be solved) but also because when
clearing the config (done on the tests) we lose keys that were added on
`load_environment` and we would need to keep track of those.