[#3196] Fall back to app_globals on Flask g
If a property is not found on `g`, check CKAN's own `app_globals`,
where stuff like `g.site_title` is kept.

The default behaviour for Flask's `g` is to raise an AttributeError if
a missing property is accessed. We have kept this behaviour here and
handled a couple of cases where `c` was accessed outside of a request.
This can be reverted to returning an empty string if necessary.