[#3196] Remove context related functions from maintain module
`deprecate_context_item` was a bit of a hack, messing with the
underlying object internal methods and hasn't really been used in ages.
It was hard to make work with the current common context object, so it
makes sense to get rid of it.

`defer_context_item` wasn't used anywhere