[#3196] New version of url_for to supports both Flask and Pylons
Essentially this calls the Flask url_for function and falls back to
Pylons if a route could not be created.

There's a lot of parameters handling before calling the Flask router
to support Pylons syntax (at least the most common cases):

    # Pylons
    url = url_for(controller='api', action='action', ver=3,
                  qualified=True)

    # Flask
    url = url_for('api.action', ver=3, _external=True)

We don't rely on the host generated by Flask, as SERVER_NAME might not
be set or might be not be up to date (as in tests changing
`ckan.site_url`). Contrary to the Routes mapper, there is no way in
Flask to pass the host explicitly, so we rebuild the URL manually
based on `ckan.site_url`, which is essentially what we alredy do on Pylons.

I experimented with automatically set SERVER_NAME based on `ckan.site_url`,
but decided against it. Apart from being magic that we want to avoid, the main
cause of concern of doing this is that when SERVER_NAME is set, the Flask
router requires the incoming URL (HTTP_HOST on environ) to have the same host
 as the one defined in SERVER_NAME, otherwise it won't match the URL and
return a 404. For example if ckan.site_url is set to http://localhost:5000
(so SERVER_NAME is localhost:5000), calls to http://127.0.0.1:5000 or
http://0.0.0.0:5000 won't work.

This would cause massive confusion to users, specially during
development or early deployment stages, so it's best to avoid it.