[#3233} Fix recline map config helpers
Previously recline_view and recline_map_view registered to separate
names for the same helper function, and the template checked for both.
But if you didn't load one of the plugins then you got a HelperError as
it hadn't been registered.

This makes all recline plugins load the same helper so it only has one
name.