Fix random order of translation files
During the generation of the JS translation files the order of the input
files previously was random by accident, as it implicitly depended on
the order of a normal Python dict. This randomly broke the ability of
extensions to override JS translations from CKAN core.

This commit fixes that issue by using an OrderedDict to ensure that
translation files provided by extensions are loaded after those of CKAN
core.