[#3365] Dispose database engine before forking a worker
rq's workers perform their tasks in separate processes. When a worker is
forked, it inherits the global state of the parent, which in our case
included SQLAlchemy's database engine and its associated connections.
The new worker process then initializes its own engine, thereby losing
the reference to its copy of the parent's engine. The latter is now not
referenced anymore and therefore eventually garbage collected, which
includes closing the associated connections. However, these connections
have been copied from the parent and may still be in use there, so
closing them in the worker leads to errors in the main process.

This commit resolves the issue by explicitly disposing the engine and
all its associated connections before the fork. This rolls back all
non-committed changes. The engine automatically re-initializes itself
once it is used again in the parent after the fork (the worker
initializes its own separate engine directly after the fork).

While this solution is clean and deterministic it still implies that you
cannot re-use session instances and ORM instances from before the fork
after the fork. However, since workers are only started in separate
paster processes and in test cases this is not an issue in normal code.