Add note about CKAN directory and personal repo
In order to help developers set up a CKAN development environment, they need to clone from a personal repository. This will make it easier to open a pull request, for developers who are not part of the CKAN organizaiton.

Also, there was a bit of ambiguity as to where to run the `pip install -e git+CKAN-repo` command. The command should be run from within `/usr/lib/ckan/default` so that subsequent steps work as expected.